+++ web/themes/custom/contribtracker/contribtracker.libraries.yml
@@ -28,6 +28,6 @@ contrib-graph:
       dist/css/components/contrib-graph.css: {}
   js:
     //cdnjs.cloudflare.com/ajax/libs/moment.js/2.4.0/moment.min.js: { external: true }
-    ts/contrib-graph.js: {}
+    js/contrib-graph.js: {}
   dependencies:
     - core/jQuery
