+++ web/themes/custom/contribtracker/templates/block/block--contribgraph.html.twig
@@ -1,37 +1,36 @@
-{% block contributed_graph_content %}
+{% block content_graph %}
+  {{ attach_library('contribtracker/contrib-graph') }}
+  <div{{ attributes }}>
+    {{ title_prefix }}
+    {% if label %}
+      <h2{{ title_attributes.addClass('nav__logo-title') }}>{{ label }}</h2>
+    {% endif %}
+    {{ title_suffix }}
+    <div class='activity-chart container'>
+      <h2 class="contrib__listing--title">{{ 'Contribution Graph' }}</h2>
+      <div class="activity-chart--wrapper">
+        <div class="activity-chart--month" id="month"></div>
 
-{{ attach_library('contribtracker/contrib-graph') }}
-<div{{ attributes }}>
-  {{ title_prefix }}
-  {% if label %}
-    <h2{{ title_attributes.addClass('nav__logo-title') }}>{{ label }}</h2>
-  {% endif %}
-  {{ title_suffix }}
-  <div class='activity-chart container'>
-    <h2 class="contrib__listing--title">{{ 'Contribution Graph' }}</h2>
-    <div class="activity-chart--wrapper">
-      <div class="activity-chart--month" id="month"></div>
+        <ol class="activity-chart--week">
+            <li>M</li>
+            <li>W</li>
+            <li>F</li>
+        </ol>
 
-      <ol class="activity-chart--week">
-          <li>M</li>
-          <li>W</li>
-          <li>F</li>
-      </ol>
-
-      <div class="activity-chart--days" id="days"></div>
-      
-      <div class="key">
-        <span>Less</span>
-        <ul class="contrib-measures activity-chart--days">
-            <li class="btn--tooltip" data-title="No Contributions" data-activity='0'></li>
-            <li class="btn--tooltip" data-title="1 Contributions" data-activity='1'></li>
-            <li class="btn--tooltip" data-title="2 Contributions" data-activity='2'></li>
-            <li class="btn--tooltip" data-title="3 Contributions" data-activity='3'></li>
-            <li class="btn--tooltip" data-title="4+ Contributions" data-activity='4'></li>
-        </ul>
-        <span>More</span>
+        <div class="activity-chart--days" id="days"></div>
+        
+        <div class="key">
+          <span>Less</span>
+          <ul class="contrib-measures activity-chart--days">
+              <li class="btn--tooltip" data-title="No Contributions" data-activity='0'></li>
+              <li class="btn--tooltip" data-title="1 Contributions" data-activity='1'></li>
+              <li class="btn--tooltip" data-title="2 Contributions" data-activity='2'></li>
+              <li class="btn--tooltip" data-title="3 Contributions" data-activity='3'></li>
+              <li class="btn--tooltip" data-title="4+ Contributions" data-activity='4'></li>
+          </ul>
+          <span>More</span>
+        </div>
       </div>
     </div>
   </div>
-</div>
 {% endblock %}
